refactor: extract review log model
